refactor(rig): C1 — the four roles get named modules behind an unchanged facade
tests/common/rig.rs -> tests/common/rig/{mod,render,materialize,invoke,
oracle}.rs. The struct, constructors, builder setters and CdcScenario
stay in mod.rs; each role module re-opens `impl Rig` (child modules see
the parent's private fields) and carries its OWN invariant in its
header doc:

- render:      pure builder-state -> YAML; never touches the fs
- materialize: file and builder never silently disagree (hand-edit guard)
- invoke:      the single execution seam the gate derives markers from
- oracle:      read-back follows the destination; missing dir != 0 parts

Every bughunt guard moved with its role, goldens intact. The gate's
marker derivation now scans the rig DIRECTORY — and its hollowing floor
fired the moment rig.rs became rig/, which is exactly the failure mode
it was RED-proven for two commits ago. No caller changed; facade
byte-identical. Gate 3/3, goldens green, scenario smokes + cli-flags
subset 5/5 live.
