+++ tests/common/rig/mod.rs
@@ -16,6 +16,11 @@ use super::runner::RIVET_BIN;
 /// stack ships. Kept as an enum so `dest_yaml` stays the single renderer for
 /// every backend (two renderers per backend is the drift the rig exists to
 /// prevent).
+mod invoke;
+mod materialize;
+mod oracle;
+mod render;
+
 enum CloudDest {
     S3 {
         bucket: String,
@@ -247,12 +252,6 @@ impl Rig {
         self
     }
 
-    pub fn out_dir(&self) -> PathBuf {
-        self.dest_override
-            .clone()
-            .unwrap_or_else(|| self.dir.path().join("out"))
-    }
-
     /// Send every export to an S3-compatible bucket instead of the rig's local
     /// tempdir — the SAME rig, so a cloud test does not fork into a hand-rolled
     /// YAML builder (the ~250 templates the rig replaced came back one engine at
@@ -297,45 +296,6 @@ impl Rig {
             prefix: prefix.to_string(),
         });
         self
-    }
-
-    /// The destination YAML for `export` — S3 when [`Rig::dest_s3`] was called,
-    /// the local tempdir otherwise. ONE renderer for both, so the primary and
-    /// secondary exports cannot drift apart (they were two `format!`s before).
-    fn dest_yaml(&self, export: &str) -> String {
-        if self.dest_stdout {
-            return "{ type: stdout }".to_string();
-        }
-        match &self.cloud_dest {
-            Some(CloudDest::S3 {
-                bucket,
-                prefix,
-                endpoint,
-            }) => format!(
-                "{{ type: s3, bucket: {bucket}, prefix: \"{prefix}/{export}/\", region: us-east-1, \
-                 endpoint: \"{endpoint}\", access_key_env: RIVET_TEST_MINIO_AK, \
-                 secret_key_env: RIVET_TEST_MINIO_SK }}"
-            ),
-            Some(CloudDest::Gcs {
-                bucket,
-                prefix,
-                endpoint,
-            }) => format!(
-                "{{ type: gcs, bucket: {bucket}, prefix: \"{prefix}/{export}/\", \
-                 endpoint: \"{endpoint}\", allow_anonymous: true }}"
-            ),
-            Some(CloudDest::Azure { container, prefix }) => format!(
-                "{{ type: azure, bucket: {container}, prefix: \"{prefix}/{export}/\", \
-                 account_name: {act}, account_key_env: RIVET_TEST_AZURITE_KEY, \
-                 endpoint: \"{ep}\" }}",
-                act = super::env::AZURITE_ACCOUNT,
-                ep = super::env::AZURITE_ENDPOINT,
-            ),
-            None => format!(
-                "{{ type: local, path: \"{}\" }}",
-                self.out_dir_for(export).display()
-            ),
-        }
     }
 
     /// Point the destination somewhere outside the rig's tempdir (e.g. a
@@ -451,14 +411,6 @@ impl Rig {
         self
     }
 
-    /// Destination directory of a named export — the primary or any secondary.
-    pub fn out_dir_for(&self, name: &str) -> PathBuf {
-        if name == self.name {
-            return self.out_dir();
-        }
-        self.dir.path().join(format!("out_{name}"))
-    }
-
     // ── THE invocation seam ─────────────────────────────────────────────
     // Every rig method that reaches the rivet binary goes through
     // `invoke_command` — one place for cross-cutting execution concerns
@@ -466,216 +418,6 @@ impl Rig {
     // the CDC conformance gate to derive its capture markers from (the
     // hand-maintained marker dictionary drifted twice in a month; see the
     // gate's `derived_capture_markers`).
-
-    /// argv for a `run` invocation: `run --config <cfg> <extra…>`.
-    fn run_argv(&self, extra: &[&str]) -> Vec<String> {
-        let cfg = self.config_path();
-        let mut v = vec![
-            "run".to_string(),
-            "--config".to_string(),
-            cfg.display().to_string(),
-        ];
-        v.extend(extra.iter().map(|a| a.to_string()));
-        v
-    }
-
-    /// argv for a non-`run` subcommand: `<args…> --config <cfg>` (clap
-    /// accepts the flag in any position).
-    fn cli_argv(&self, args: &[&str]) -> Vec<String> {
-        let cfg = self.config_path();
-        let mut v: Vec<String> = args.iter().map(|a| a.to_string()).collect();
-        v.push("--config".to_string());
-        v.push(cfg.display().to_string());
-        v
-    }
-
-    /// The single `Command` builder behind every runner wrapper.
-    fn invoke_command(&self, argv: &[String], envs: &[(&str, &str)]) -> std::process::Command {
-        let mut cmd = std::process::Command::new(super::runner::RIVET_BIN);
-        cmd.args(argv);
-        for (k, v) in envs {
-            cmd.env(k, v);
-        }
-        cmd
-    }
-
-    /// Run to completion and collect the output.
-    fn invoke(&self, argv: &[String], envs: &[(&str, &str)]) -> std::process::Output {
-        self.invoke_command(argv, envs)
-            .output()
-            .expect("spawn rivet binary")
-    }
-
-    /// Run an ARBITRARY subcommand against this rig's config: `rivet <args…>
-    /// --config <cfg>`.
-    ///
-    /// NOT for `apply`, which takes a PLAN PATH rather than `--config` — that is
-    /// [`Rig::apply_env`]'s job. This method appends the config flag, so it
-    /// fits the subcommands that read one: `plan`, `check`, `validate`, `doctor`.
-    ///
-    /// `run_args`/`run_args_env` hard-code the `run` subcommand, so a test for
-    /// `check`, `validate`, `doctor` or `init` had no way through the rig and
-    /// dropped to a raw `Command`. The config flag is appended, which clap
-    /// accepts in any position.
-    pub fn cli(&self, args: &[&str]) -> std::process::Output {
-        self.cli_env(args, &[])
-    }
-
-    /// [`Rig::cli`] with environment variables — needed wherever the config
-    /// declares `url_env:`, since the process must be able to resolve it.
-    pub fn cli_env(&self, args: &[&str], envs: &[(&str, &str)]) -> std::process::Output {
-        self.invoke(&self.cli_argv(args), envs)
-    }
-
-    /// `rivet plan --export <this rig's export> --format json --output <out>`,
-    /// plus `extra` args (`--param k=v`, `--annotate-waves`, …).
-    ///
-    /// The plan→apply pair is the one CLI flow whose two halves take DIFFERENT
-    /// subjects — `plan` reads the config, `apply` reads the artifact — so a
-    /// test that wants the round trip had to spell the six-flag `plan`
-    /// invocation itself and then drop out of the rig entirely for `apply`
-    /// (every call site in `live_plan_apply.rs` does exactly that). The export
-    /// name comes from the rig rather than the caller, which is what keeps the
-    /// artifact, the destination and the `export_metrics` rows talking about the
-    /// same export.
-    pub fn plan_json_env(
-        &self,
-        out: &Path,
-        extra: &[&str],
-        envs: &[(&str, &str)],
-    ) -> std::process::Output {
-        let out = out.to_str().expect("plan output path must be utf-8");
-        let mut args: Vec<&str> = vec![
-            "plan",
-            "--export",
-            self.name.as_str(),
-            "--format",
-            "json",
-            "--output",
-            out,
-        ];
-        args.extend_from_slice(extra);
-        self.cli_env(&args, envs)
-    }
-
-    /// `rivet apply <plan.json>` plus `extra` args (`--force`, `--resume`), with
-    /// `envs` set — the counterpart of [`Rig::plan_json_env`].
-    ///
-    /// The ONE subcommand that takes a PLAN PATH instead of `--config`, which is
-    /// why it cannot go through [`Rig::cli_env`] (that appends `--config`) and
-    /// why it needs its own method rather than a raw `Command`. It still belongs
-    /// on the rig: `apply` writes into the rig's destination and opens
-    /// `.rivet_state.db` next to the rig's CONFIG (the artifact records the
-    /// config path), so the read-backs a test does afterwards — `out_dir()`,
-    /// the state DB — are the rig's, not the plan file's.
-    ///
-    /// `envs` is not optional in practice: a plan/apply round trip needs
-    /// [`Rig::source_url_env`] (an inline URL is redacted into the artifact and
-    /// apply then cannot reconnect), so the variable must be set on BOTH legs.
-    pub fn apply_env(
-        &self,
-        plan: &Path,
-        extra: &[&str],
-        envs: &[(&str, &str)],
-    ) -> std::process::Output {
-        let mut args: Vec<&str> = vec!["apply", plan.to_str().expect("plan path must be utf-8")];
-        args.extend_from_slice(extra);
-        super::runner::run_rivet_env(&args, envs)
-    }
-
-    /// Spawn `rivet run` and hand back the LIVE child, output discarded.
-    ///
-    /// For tests that must act on a running process — signal it, inspect its
-    /// children, watch the staged `.tmp` appear — rather than wait for an exit
-    /// status. `run_args_env` blocks until completion and so cannot express them.
-    /// The caller owns the `Child` and must reap it.
-    pub fn spawn_args_env(&self, extra: &[&str], envs: &[(&str, &str)]) -> std::process::Child {
-        self.invoke_command(&self.run_argv(extra), envs)
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .spawn()
-            .expect("spawn rivet")
-    }
-
-    /// Run `rivet run --config <rig cfg>` plus `extra` args, with `envs` set.
-    ///
-    /// The affordance the crash-recovery files were bypassing the rig for: they
-    /// built their YAML through `Rig` and then dropped to a raw
-    /// `Command::new(RIVET_BIN)` because the rig could express an env var OR a
-    /// config, never extra ARGS (`--export`, `--resume`) alongside a fault
-    /// injection. That one gap accounted for most of the hand-rolled invocations
-    /// in `live_chunked_recovery.rs` and its siblings.
-    pub fn run_args_env(&self, extra: &[&str], envs: &[(&str, &str)]) -> std::process::Output {
-        self.invoke(&self.run_argv(extra), envs)
-    }
-
-    /// `run_args_env` with no env — a plain run with extra flags.
-    pub fn run_args(&self, extra: &[&str]) -> std::process::Output {
-        self.run_args_env(extra, &[])
-    }
-
-    /// Run with an extra environment variable (fault injection); returns the
-    /// raw output — the caller asserts success or failure.
-    pub fn run_with_env(&self, key: &str, val: &str) -> std::process::Output {
-        self.run_args_env(&[], &[(key, val)])
-    }
-
-    /// Run with SEVERAL extra environment variables (e.g. RIVET_STATE_URL to pick the
-    /// Postgres state backend AND RIVET_TEST_PANIC_AT to inject a crash in one run);
-    /// returns the raw output — the caller asserts success or failure.
-    pub fn run_with_envs(&self, envs: &[(&str, &str)]) -> std::process::Output {
-        self.run_args_env(&[], envs)
-    }
-
-    /// [`Rig::run_with_envs`] under a WALL-CLOCK CEILING — `None` if the child
-    /// had to be killed.
-    ///
-    /// `run_with_envs` bottoms out in `Command::output()`, which blocks with no
-    /// timeout. That is fine for a test whose failure mode is a wrong value, and
-    /// wrong for one whose failure mode is a HANG: the governor deadlock
-    /// (`governor_does_not_deadlock_when_chunks_fail`) is a live regression
-    /// class, and a test that hangs while holding `quiet_window_guard` converts
-    /// one red test into an indefinite stall of every test that takes the same
-    /// cross-process lock — plus, for the pressure tests, a background writer
-    /// that keeps hammering the shared server forever.
-    ///
-    /// stdout/stderr go to FILES rather than pipes: polling `try_wait` while a
-    /// child fills a pipe buffer nobody drains is its own deadlock (the reason
-    /// the hand-rolled watchdogs in `live_governor.rs` redirect to a file).
-    pub fn run_with_envs_bounded(
-        &self,
-        envs: &[(&str, &str)],
-        timeout: std::time::Duration,
-    ) -> Option<std::process::Output> {
-        let out_path = self.dir.path().join("bounded.stdout");
-        let err_path = self.dir.path().join("bounded.stderr");
-        let mut cmd = self.invoke_command(&self.run_argv(&[]), envs);
-        cmd.stdout(std::fs::File::create(&out_path).expect("bounded stdout file"))
-            .stderr(std::fs::File::create(&err_path).expect("bounded stderr file"));
-        let mut child = cmd.spawn().expect("spawn rivet binary");
-        let start = std::time::Instant::now();
-        loop {
-            if let Some(status) = child.try_wait().expect("try_wait rivet") {
-                return Some(std::process::Output {
-                    status,
-                    stdout: std::fs::read(&out_path).unwrap_or_default(),
-                    stderr: std::fs::read(&err_path).unwrap_or_default(),
-                });
-            }
-            if start.elapsed() >= timeout {
-                let _ = child.kill();
-                let _ = child.wait();
-                return None;
-            }
-            std::thread::sleep(std::time::Duration::from_millis(50));
-        }
-    }
-
-    pub fn checkpoint(&self) -> PathBuf {
-        self.ckpt_override
-            .clone()
-            .unwrap_or_else(|| self.dir.path().join("cdc.ckpt"))
-    }
 
     /// Override the checkpoint path (resume/crash suites share one
     /// checkpoint across several configs — the rig renders, the test owns
@@ -698,238 +440,6 @@ impl Rig {
     pub fn with_format(mut self, fmt: &'static str) -> Self {
         self.format = fmt;
         self
-    }
-
-    /// The rendered YAML — for suites that own their config-file lifetime.
-    pub fn yaml(&self) -> String {
-        self.render()
-    }
-
-    /// Materialized config path — for bespoke invocations (`validate`,
-    /// custom envs) the rig doesn't wrap.
-    /// The export name this rig rendered into the config — the key
-    /// `export_metrics` rows are stored under, so a test can read the run's own
-    /// metrics without re-deriving the name it did not choose.
-    pub fn export_name(&self) -> &str {
-        &self.name
-    }
-
-    /// Re-render THIS rig's config with extra export lines appended, over the
-    /// SAME config path (and therefore the same adjacent `.rivet_state.db`).
-    ///
-    /// For two-phase fixtures whose subject is a SECOND run under a CHANGED
-    /// config against the FIRST run's state — e.g. gremlin G2: run 1 advances an
-    /// incremental cursor with no quality gate, run 2 adds `row_count_min` and
-    /// must see the gate fire on the exhausted cursor. Without this the test
-    /// hand-writes a second YAML into the rig's dir (`write_config`), which is
-    /// the per-file config builder the rig exists to replace.
-    /// Replace the first export line starting with `prefix` — for two-phase
-    /// fixtures whose SECOND run must change a knob (`parallel: 1` -> `2`)
-    /// over the SAME config path and state DB. Panics if no line matches:
-    /// silently not-replacing is how a resume leg measures nothing.
-    pub fn replace_export_line(&mut self, prefix: &str, new_line: &str) -> PathBuf {
-        let slot = self
-            .extra_lines
-            .iter_mut()
-            .find(|l| l.starts_with(prefix))
-            .unwrap_or_else(|| {
-                panic!("replace_export_line: no export line starts with {prefix:?}")
-            });
-        *slot = new_line.to_string();
-        let path = self.dir.path().join("rig.yaml");
-        std::fs::write(&path, self.render()).expect("re-render rig config");
-        path
-    }
-
-    pub fn amend_export_lines(&mut self, lines: &[&str]) -> PathBuf {
-        for l in lines {
-            self.extra_lines.push((*l).to_string());
-        }
-        // Same filename and same renderer as config_path — a divergent copy
-        // here would run yesterday's config while looking amended.
-        let path = self.dir.path().join("rig.yaml");
-        std::fs::write(&path, self.render()).expect("re-render rig config");
-        path
-    }
-
-    /// Materialize this rig's config into a CALLER-owned directory and return
-    /// the path — for scenarios where the config must OUTLIVE the rig value.
-    ///
-    /// This is the sanctioned answer to the temporary-rig-drop trap: a
-    /// `rig.config_path()` handed out of a helper dies with the rig's tempdir
-    /// (live_cdc documented the trap and kept `write_config(d, &rig.yaml())`
-    /// round-trips as the workaround; live_cdc_mssql hit it live during the
-    /// migration). Destinations are still pre-created exactly as
-    /// [`Rig::config_path`] does, so the two materializations cannot drift.
-    pub fn config_in(&self, dir: &std::path::Path) -> PathBuf {
-        if self.dest_precreate {
-            std::fs::create_dir_all(self.out_dir()).unwrap();
-        }
-        for e in &self.extra_exports {
-            std::fs::create_dir_all(self.out_dir_for(&e.name)).unwrap();
-        }
-        let cfg = dir.join("rig.yaml");
-        std::fs::write(&cfg, self.render()).unwrap();
-        cfg
-    }
-
-    pub fn config_path(&self) -> PathBuf {
-        // Materialization point: the ONLY place the rig touches the
-        // filesystem (yaml()/render() stay pure — the offline goldens were
-        // mkdir-ing /tmp/o as a side effect of rendering a string).
-        if self.dest_precreate {
-            std::fs::create_dir_all(self.out_dir()).unwrap();
-        }
-        for e in &self.extra_exports {
-            std::fs::create_dir_all(self.out_dir_for(&e.name)).unwrap();
-        }
-        let cfg = self.dir.path().join("rig.yaml");
-        let rendered = self.render();
-        // Every run helper re-materializes through here, so a caller who
-        // HAND-EDITED the file would have the patch silently clobbered and the
-        // test would run the un-patched config while looking patched (bughunt:
-        // three chunked-recovery resume tests measured nothing on their
-        // changed-parallelism leg exactly this way). Refuse loudly instead:
-        // the sanctioned ways to change a materialized config are
-        // [`Rig::amend_export_lines`] and [`Rig::replace_export_line`].
-        if let Ok(existing) = std::fs::read_to_string(&cfg)
-            && existing != rendered
-        {
-            panic!(
-                "rig.yaml at {} was edited outside the rig; every rig run \
-                 re-renders the config, so the edit would be silently \
-                 discarded. Mutate the rig instead (amend_export_lines / \
-                 replace_export_line), or run the edited file via run_rivet.",
-                cfg.display()
-            );
-        }
-        std::fs::write(&cfg, rendered).unwrap();
-        cfg
-    }
-
-    fn render(&self) -> String {
-        let tables = match &self.query {
-            Some(q) => format!("query: \"{q}\""),
-            None if self.tables.len() == 1 => format!("table: {}", self.tables[0]),
-            None => format!("tables: [{}]", self.tables.join(", ")),
-        };
-        let cdc_lines: Vec<String> = self
-            .cdc_lines
-            .iter()
-            .map(|l| {
-                if l == "__CKPT__" {
-                    format!("checkpoint: \"{}\"", self.checkpoint().display())
-                } else {
-                    l.clone()
-                }
-            })
-            .collect();
-        let cdc = if cdc_lines.is_empty() {
-            String::new()
-        } else {
-            format!("    cdc: {{ {} }}\n", cdc_lines.join(", "))
-        };
-        let extra: String = self
-            .extra_lines
-            .iter()
-            .map(|l| format!("    {l}\n"))
-            .collect();
-        let source = if self.source_lines.is_empty() {
-            match &self.url_env {
-                Some(v) => format!("source: {{ type: {}, url_env: {v} }}", self.source_type),
-                None => format!(
-                    "source: {{ type: {}, url: \"{}\" }}",
-                    self.source_type, self.source_url
-                ),
-            }
-        } else {
-            let extra: String = self
-                .source_lines
-                .iter()
-                .map(|l| format!("  {l}\n"))
-                .collect();
-            match &self.url_env {
-                Some(v) => format!(
-                    "source:\n  type: {}\n  url_env: {v}\n{extra}",
-                    self.source_type
-                ),
-                None => format!(
-                    "source:\n  type: {}\n  url: \"{}\"\n{extra}",
-                    self.source_type, self.source_url
-                ),
-            }
-            .trim_end()
-            .to_string()
-        };
-        // Secondary exports (see `Rig::also_export`) each get their OWN
-        // destination, which is what the multi-export configs under test
-        // actually declare — per-export failure isolation is only observable
-        // when the outputs are separable.
-        let secondaries: String = self
-            .extra_exports
-            .iter()
-            .map(|e| {
-                let lines: String = e.lines.iter().map(|l| format!("    {l}\n")).collect();
-                let subject = match &e.table {
-                    Some(t) => format!("table: {t}"),
-                    None => format!("query: \"{}\"", e.query),
-                };
-                let cdc = if e.cdc_lines.is_empty() {
-                    String::new()
-                } else {
-                    format!("    cdc: {{ {} }}\n", e.cdc_lines.join(", "))
-                };
-                format!(
-                    "  - name: {n}\n    {subject}\n    mode: {m}\n    format: {f}\n{cdc}{lines}    destination: {o}\n",
-                    n = e.name,
-                    m = e.mode,
-                    f = self.format,
-                    o = self.dest_yaml(&e.name),
-                )
-            })
-            .collect();
-        let yaml = format!(
-            "{source}\nexports:\n  - name: {name}\n    {tables}\n    mode: {mode}\n    format: {fmt}\n{cdc}{extra}    destination: {out}\n{secondaries}",
-            name = self.name,
-            tables = tables,
-            mode = self.mode,
-            fmt = self.format,
-            out = self.dest_yaml(&self.name.clone()),
-        );
-        yaml
-    }
-
-    /// Run rivet; panic unless it succeeds.
-    pub fn run_ok(&self) {
-        let out = self.run_args(&[]);
-        assert!(
-            out.status.success(),
-            "rig run failed for '{}':\n{}",
-            self.name,
-            String::from_utf8_lossy(&out.stderr)
-        );
-    }
-
-    /// Run rivet expecting a loud failure; returns combined output.
-    pub fn run_expect_fail(&self) -> String {
-        let out = self.run_args(&[]);
-        assert!(
-            !out.status.success(),
-            "rig run for '{}' was expected to fail",
-            self.name
-        );
-        format!(
-            "{}{}",
-            String::from_utf8_lossy(&out.stdout),
-            String::from_utf8_lossy(&out.stderr)
-        )
-    }
-
-    /// Run rivet; return the raw output without asserting either way — for tests
-    /// whose VALID outcomes include both success and a loud failure (e.g. a
-    /// mid-stream outage that rivet may either retry through or safely refuse).
-    pub fn run(&self) -> std::process::Output {
-        self.run_args(&[])
     }
 
     /// Put the destination under the shared bind mount so the DuckDB validator
@@ -974,29 +484,6 @@ impl Rig {
         self.oracle_container_dir
             .as_deref()
             .expect("oracle_dir needs Rig::duckdb_oracle() at construction")
-    }
-
-    /// Run and read every parquet part back — the canonical
-    /// capture-and-verify shape the outcome gate keys on.
-    ///
-    /// CUMULATIVE: reads the whole out dir, so a second call after more runs
-    /// returns old parts too — poll loops must ASSIGN the result, never `+=`
-    /// it (bughunt: a `rows +=` poll double-counted one part and masked a
-    /// dropped event).
-    ///
-    /// Refuses cloud/stdout rigs loudly: the data lands in the bucket (or on
-    /// stdout), the local out dir is empty by construction, and the returned
-    /// `[]` is indistinguishable from "the run wrote zero rows" — every
-    /// negative assertion downstream would be vacuous forever.
-    pub fn run_and_read(&self) -> Vec<arrow::record_batch::RecordBatch> {
-        assert!(
-            self.cloud_dest.is_none() && !self.dest_stdout,
-            "run_and_read reads the rig's LOCAL out dir, but this rig writes to \
-             a cloud/stdout destination — read the destination store directly \
-             (mc/HTTP oracle), or drop the cloud dest"
-        );
-        self.run_ok();
-        read_all_parts(&self.out_dir())
     }
 }
